OPTIMIZATION: Reduce CPU usage by ~6%
This fixes that the CPU is put on full constraint, even with an FPS cap active.
It happens because the accurate "sleeping" happens with the Sleep(0) (this doesn't actually sleep, at best it tells the cpu scheduling to reschedule execution for other threads) in the fpsTimer while loop.

But there should also be an inaccurate proper sleep to avoid maxing out one core, 1ms is good.